pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus Vidal",email = "matheusvidal140@gmail.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest --cov=fast_zero -vv'
post_test = 'coverage html'
lint = 'ruff check . ; ruff check . --diff'
format = 'ruff check . --fix ; ruff format .'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'